:sparkles: Fixed the Pick-Up Time
The pick-up time box now uses the proper microinteractions that denote when a time is selected or not.

Like the tip selectors, it's essentially a glorified radio button selector, so we would have to have some values pre-written into the options for JS or PHP to actually use them

For the purposes of a demo, however, they should do fine.